+++ src/compas/geometry/rotation.py
@@ -10,6 +10,7 @@ following online resources:
 Many thanks to Christoph Gohlke, Martin John Baker, Sachin Joglekar and Andrew
 Ippoliti for providing code and documentation.
 """
+
 from compas.utilities import flatten
 from compas.geometry import normalize_vector
 from compas.geometry import cross_vectors