+++ src/compas_ghpython/components/Compas_Info/code.py
@@ -1,6 +1,7 @@
 """
 Displays information about the active COMPAS environment.
 """
+
 import compas_bootstrapper
 from ghpythonlib.componentbase import executingcomponent as component